better

+++ src/utils/location.hpp
@@ -5,9 +5,9 @@
 struct Location {
   Location();
 
-  int lineNumber_;
-  int firstColumn_;
-  int lastColumn_;
+  size_t lineNumber_;
+  size_t firstColumn_;
+  size_t lastColumn_;
 };
 
 inline Location::Location() : lineNumber_{1}, firstColumn_{1}, lastColumn_{1} {}
